requirements.txt
PySide6>=6.6.1
pandas>=2.1.4
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.6.1
numpy>=1.26.0
openpyxl>=3.1.2
PyJWT>=2.8.0
aiohttp>=3.9.1
tqdm>=4.66.1
pyinstaller>=6.5.0 
selenium>=4.0.0 
selenium-wire>=5.1.0
blinker==1.4 